// src/education/users/item/assignments/item/submissions/item/resources/item/educationSubmissionResourceItemRequestBuilderDeleteRequestConfiguration.ts
import {RequestOption} from '@microsoft/kiota-abstractions';

export interface EducationSubmissionResourceItemRequestBuilderDeleteRequestConfiguration {
    /** Request headers */
    headers?: Record<string, string> | undefined;
    /** Request options */
    options?: RequestOption[] | undefined;
}